requirements.txt
fastapi
uvicorn
langchain
chromadb
sentence-transformers
pydantic
python-dotenv
pyyaml
uvicorn[standard]
black
flake8
pytest
langchain_core
langchain_community
torch
transformers
